@@ -14,14 +14,20 @@ classifiers = [
     ]
 readme = "README.md"
 
+# This build system is required so that falcon 3.1.1 is used as a subdependency 
+# of morph-kgc, otherwise building morph-kgc will fail
+[build-system]
+requires = ["poetry-core>=1.1.10"]
+build-backend = "poetry.core.masonry.api"
+
 [tool.poetry.dependencies]
-python = ">=3.9,<=3.13"
+python = "^3.13.0"
 bs4 = "^0.0.1"
 click = "^8.1.7"
 click-option-group = "^0.5.6"
 matplotlib = "^3.8.2"
 nltk = "^3.9.0"
-numpy = "^1.26.3"
+numpy = "^2.3.0"
 pandas = "^2.1.4"
 rdflib = "^7.0.0"
 textblob = "^0.17.1"
@@ -33,13 +39,14 @@ contractions = "^0.1.73"
 chardet = "^5.2.0"
 imbalanced-learn = "^0.11.0"
 pytest = "^7.4.4"
-morph-kgc = "^2.6.4"
+falcon = "^4.0.0"
+morph-kgc = "^2.8.1"
 bibtexparser = "^1.4.1"
 nbformat = "^5.9.2"
 markdown = "^3.5.2"
 rdflib-jsonld = "^0.6.2"
 requests = "^2.31.0"
-scikit-learn = "1.3.2"
+scikit-learn = "^1.4.2"
 pyyaml = "^6.0.2"
 lxml = "^5.1.0"
 
@@ -48,7 +55,3 @@ somef = "somef.__main__:cli"
 
 [tool.poetry.urls]
 homepage = "https://github.com/KnowledgeCaptureAndDiscovery/somef"
-
-[build-system]
-requires = ["poetry-core>=1.1.10"]
-build-backend = "poetry.core.masonry.api"
